+++ src/nscheme.h
@@ -25,7 +25,7 @@
 
 typedef struct {
     int nodes[3];
-    double matriz[6]; // C11 C22 C33 C12 C13 C23
+    float matriz[6]; // C11 C22 C33 C12 C13 C23
 } elementri;
 
 typedef struct {
